Add validation for required --net-bridged parameter
Updated the Run.swift file in the tart command directory to include validation for the --net-bridged parameter when executing the tart command. This has been implemented to accommodate the McKinsey specific version of Tart where NAT networking is not allowed. The command will now throw a validation error if the bridged network interface is not specified.