Refactor netBridged validation in Run.swift
Updated the condition for checking if netBridged parameter is defined. Previously, an equality operator was used to check if the count of netBridged is zero. This has been streamlined to use the more concise and idiomatic Swift isEmpty property, increasing code clarity and maintainability.

Furthermore, the error message has been improved to be more direct. The old verbose error message has been replaced with a shorter and clearer message.